add redis db and created reset-password endpoint
