+++ Pipfile
@@ -6,7 +6,7 @@ name = "pypi"
 [packages]
 alembic = "==1.9.0"
 asyncpg = "==0.27.0"
-aioredis = "*"
+redis = "*"
 pytest = "==7.2.0"
 pydantic = "==1.10.11"
 psycopg2-binary = "==2.9.5"
@@ -24,4 +24,4 @@ sqlalchemy = "==2.0.0"
 [dev-packages]
 
 [requires]
-python_version = "3.11"
+python_version = "3.10.6"
