+++ src/api/actions/auth.py
@@ -1,14 +1,7 @@
-import os
-import random
-import string
-
-import boto3
 from fastapi.security import OAuth2PasswordBearer
 from fastapi_jwt_auth import AuthJWT
-from pydantic import EmailStr
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from settings import localstack_endpoint_url
 from src.api.schemas import Settings
 from src.db.dals import UserDAL
 from src.db.models import User
@@ -34,29 +27,6 @@ async def authenticate_user(
     if not Hasher.verify_password(password, user.hashed_password):
         return
     return user
-
-
-def generate_reset_password_token(length: int = 32):
-    characters = string.ascii_letters + string.digits
-    return "".join(random.choice(characters) for _ in range(length))
-
-
-async def send_reset_token(email: str, reset_token):
-    ses_client = boto3.client(
-        "ses",
-        region_name="us-east-1",
-        endpoint_url=localstack_endpoint_url,
-    )
-
-    subject = "Reset Password"
-    body = f"Password reset link: http://localhost:8000/auth/reset-password?token={reset_token}"
-
-    response = ses_client.send_email(
-        Source="sender@example.com",
-        Destination={"ToAddresses": [email]},
-        Message={"Subject": {"Data": subject}, "Body": {"Text": {"Data": body}}},
-    )
-    return response
 
 
 @AuthJWT.load_config
